Cambio de tipo float a double

+++ calculadora/src/calculadora/Calculadora.java
@@ -2,7 +2,7 @@ package calculadora;
 
 public class Calculadora {
 	//atributos
-	float num1, num2;
+	double num1, num2;
 	double resultado;
 	
 	//constructor
